Fix Error reported previously; Add TextFile example
A bug reported on commit 8647232 was fixed. Need to check why it was
there and if any future changes need to be made.

If examined, the mydrive data base shows that all objects are linked
and they are being created correctly.

Constructors done and tested manually, subject to little changes. closes #2