Fix bug in createIntermediatePath()
This method did not previously consider the fact that the path might not begin with / or ./
(it can simply be the name of a directory starting in currentDir). That is now fixed.

The method createDirectoryFromPath() was also updated, now covering the fact that the path
that leads to a directory might end in / . It now works around that and accepts both cases, with or without a final / .