Minor updates to domain + recursive file removal
Deleted unnecessary constructors for AbstractFile.
Deleted some redundant code lines.
Updated exceptions thrown by some methods, and implemented the recursive
file removal - when trying to delete a Directory, it will not throw
an exception when it is not empty, but rather try to delete everything
inside (at any depth) if is it possible to do so.