Update Variable class
Was not connecting a Variable to a Login. In order to accomodate this
requirement, and to support the variable-adding service, the Variable
constructor was made protected and a Variable can only be added through
an instance of the Login class - if the Variable already exists in the
Login instance with the name of the Variable one is trying to add, its
value will be replaced by the new one.